requirements_pro.txt
# Flask Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# Database
SQLAlchemy==2.0.21
Alembic==1.12.0

# Security
Flask-Bcrypt==1.0.1
cryptography==41.0.4
PyJWT==2.8.0
bleach==6.0.0

# Internationalization
Flask-Babel==3.1.0
Babel==2.12.1

# Performance & Caching
Flask-Compress==1.13
Flask-Caching==2.1.0
redis==5.0.0

# Rate Limiting
Flask-Limiter==3.5.0

# Email
Flask-Mail==0.9.1

# File Handling
Pillow==10.0.0
openpyxl==3.1.2
pandas==2.1.1

# HTTP Requests
requests==2.31.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Two-Factor Authentication
pyotp==2.9.0
qrcode==7.4.2

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0

# Development Tools
python-dotenv==1.0.0

# Production Server
waitress==2.1.2
gunicorn==21.2.0

# Monitoring & Logging
psutil==5.9.5

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# API Documentation
flask-restx==1.2.0

# Background Tasks
celery==5.3.2

# Database Drivers
psycopg2-binary==2.9.7
PyMySQL==1.1.0

# Charts & Visualization
matplotlib==3.7.2
seaborn==0.12.2

# PDF Generation
reportlab==4.0.4
weasyprint==59.0

# Excel Export
xlsxwriter==3.1.3

# Configuration Management
python-decouple==3.8

# Health Checks
flask-healthz==0.0.3

# CORS Support
Flask-CORS==4.0.0

# Session Management
Flask-Session==0.5.0

# Form Validation
email-validator==2.0.0

# JSON Web Tokens
authlib==1.2.1

# Password Strength
zxcvbn==4.4.28

# HTML Parsing
beautifulsoup4==4.12.2

# CSV Handling
csvkit==1.1.1

# Backup Scheduling
APScheduler==3.10.4

# Memory Profiling
memory-profiler==0.61.0

# Code Quality
flake8==6.0.0
black==23.7.0
isort==5.12.0

# Documentation
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Async Support
Flask-SocketIO==5.3.6
eventlet==0.33.3

# Database Migration
Flask-Migrate==4.0.5

# Admin Interface
Flask-Admin==1.6.1

# Data Serialization
msgpack==1.0.5

# Compression
brotli==1.0.9

# User Agent Parsing
user-agents==2.2.0

# Phone Number Validation
phonenumbers==8.13.19

# Country Data
pycountry==22.3.5

# Language Detection
langdetect==1.0.9

# Statistics
scipy==1.11.2
numpy==1.25.2

# API Client
httpx==0.24.1

# Caching
diskcache==5.6.3

# Debugging
flask-debugtoolbar==0.13.1

# SSL/TLS
pyOpenSSL==23.2.0

# Network Security
dnspython==2.4.2

# File Watching
watchdog==3.0.0

# Cloud Storage
boto3==1.28.57
google-cloud-storage==2.10.0

# Social Authentication
Flask-Dance==7.0.0

# OAuth
oauthlib==3.2.2

# Metrics Collection
prometheus-client==0.17.1

# Application Performance Monitoring
sentry-sdk==1.32.0

# Real-time Communication
pusher==3.3.2

# WebSocket Support
websockets==11.0.3

# Search Engine
elasticsearch==8.9.0
whoosh==2.7.4

# Machine Translation
googletrans==4.0.0rc1

# Barcode Generation
python-barcode==0.14.0

# QR Code Generation
segno==1.5.2

# Calendar Support
icalendar==5.0.7

# Task Scheduling
schedule==1.2.0

# Log Analysis
loguru==0.7.0

# Configuration Validation
schema==0.7.5

# API Versioning
flask-restful==0.3.10

# Database Pooling
SQLAlchemy-Utils==0.41.1

# Development Server
flask-shell-ipython==1.4.0
